Fix palooza (desc.)
Textures are now enabled before use and disabled after. (enabled when
needed really)
Added alphafunc to sort of fix the text glyphs (it's supposed to be
temporary but I'm happy enough with it to rest it now.
QuickFont depth change update.
RectangleF.IntersectsWith does not work for negative widths/heights,
made the rectangles for viewdraw-checking always start in the top left
corner.