A lot of shiet incl. a ShapesIntersecting rework
Added a new step timing method which is the default, and a way to set it
back to catch-up mode.
Removed Collision precision workaround that didn't work around an awful
lot.
Made Input button checks decide what is going to be checked by looking
at the button.
Fixed (-Is)GamepadConnected not checking the right gamepad.
ShapesIntersecting is now changed to an IShape extension method which
makes it clearer and more logical to work with.
The intersection functions are now more versatile in that they handle
checks for both moving and static objects.
Implemented the missing checks for AABBs and Points.
Collisions against objects that can't move anymore can still occur, duh.