GUITools
Button and Scrollbar. Button replaces ButtonObject and can be customized
and created without needing to be inherited which is way easier.
Scrollbar is still very experimental.
Made drawing keep the order after sorting by depth so what's added first
on the same depth will be drawn behind.
Fixed sprites not indexing right when the total horizontal size couldn't
be integer-divided by the frame size.
Added way to get a product of the X and Y of a Point easier.