Made View way more complicated, and useful I guess
View can now have absolute and relative values for sizes of both areas
and viewports, and also for positions of viewports. It can also have
negative values which are meant to fill (a factor of) the remaining
space of the window. This will also be a factor of the window space left
for negative area size values.
Fixed FilledRects not taking into account the view's offset.